code/requirements.txt
flask
flask-caching
python-dotenv
pandas
psycopg2-binary
requests
sqlalchemy
